Moves common File interface and classes
