+++ basyx.submodelrepository/basyx.submodelrepository-backend/src/main/java/org/eclipse/digitaltwin/basyx/submodelrepository/backend/SubmodelBackendProvider.java
@@ -26,6 +26,7 @@
 package org.eclipse.digitaltwin.basyx.submodelrepository.backend;
 
 import org.eclipse.digitaltwin.aas4j.v3.model.Submodel;
+import org.eclipse.digitaltwin.basyx.core.file.FileRepository;
 import org.springframework.data.repository.CrudRepository;
 
 /**
